bashrc: path_append for PATH and PYTHONPATH
Don't append a path to PATH or PYTHONPATH that already exists on it.
The function `path_append` iterates through the segments on the path and
checks them against the path to append. If no match is made, then append
the path to the existing path.

The new path is echoed back to be exported.