deploy: install Dropbox CLI, use it to install Dropbox and start daemon
Handily, I found that the Dropbox CLI can install Dropbox for you when you
try to start the daemon. It requires a few modifications to the script
because it would normally prompt for "Are you sure you want to install?".
I just remove the lines.

I also fix their "running" command which returns 1 for True (is running)
and 0 for False (is not running), which follows Python's interpretation
of True/False as integers but not Unix-style exit codes.